requirements.txt
bcrypt
tortoise-orm
fastapi
uvicorn
pydantic
tortoise-orm[asyncpg]
pytest
pytest-asyncio
validators
python-dotenv
aerich